# openapi.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Search Courses API",
    "description": "API for searching courses from various sites.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local server"
    }
  ],
  "paths": {
    "/api/v2/post": {
      "get": {
        "tags": [
          "post-controller"
        ],
        "summary": "Busca posts",
        "description": "Retorna uma lista de posts baseados em filtros opcionais",
        "operationId": "findAll",
        "parameters": [
          {
            "name": "content",
            "in": "query",
            "description": "Texto para buscar no título ou resumo do post",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Erro no servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPostResponse"
                }
              }
            }
          },
          "200": {
            "description": "Lista de posts retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostEntity"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/post/{id}/click": {
      "get": {
        "tags": [
          "post-controller"
        ],
        "summary": "Registra clique no post",
        "description": "Este endpoint registra um clique para o post especificado.\n**Nota:** Este endpoint realiza alterações no banco de dados (incrementa o contador de cliques), o que é um efeito colateral.\n",
        "operationId": "registerClick",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do post para registrar o clique",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Redirecionamento para a URL do post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdClickUrlResponse"
                }
              }
            }
          },
          "400": {
            "description": "ID inválido ou erro na requisição",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorClickPostsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/post/click/counts": {
      "get": {
        "tags": [
          "post-controller"
        ],
        "summary": "Retorna contagem de cliques por post",
        "description": "Retorna um array com a contagem de cliques organizada por post.\nCada item do array contém o título do post como chave e um objeto com:\n- count: número total de cliques\n- date: data do último registro\n",
        "operationId": "getClickCounts",
        "responses": {
          "200": {
            "description": "Contagens de cliques obtidas com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClickCountsUrlResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao obter contagens de cliques",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorClickPostsResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorPostResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "PostEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "site": {
            "$ref": "#/components/schemas/SiteEntity"
          },
          "indexDate": {
            "type": "string",
            "example": "2025-05-14 10:31:03.803860"
          },
          "pubDate": {
            "type": "string",
            "example": "2025-05-14 10:31:03.803860"
          }
        }
      },
      "SiteEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "rss": {
            "type": "string"
          },
          "about": {
            "type": "string"
          },
          "iframe": {
            "type": "boolean"
          }
        }
      },
      "IdClickUrlResponse": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ErrorClickPostsResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ClickCountsUrlResponse": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "count": {
            "type": "string"
          }
        }
      }
    }
  }
}